feat(Compiler): Pull out implementation of GeneratedInjector calls to methods.
This allows progress towards https://github.com/dart-lang/angular/issues/772,
where multi-tokens/providers are going to need to aggregate results of multiple
different tokens for the given injector, i.e.:

if (identical(token, someMultiToken)) {
  return [_get1(), _get2(), _get3()];
}

This shouldn't affect Dart2JS performance assuming we can inline just fine. It
will be possible to automatically inline in the GeneratedInjector when we know
that no providers are using multi, but it is probably not worth it yet.

PiperOrigin-RevId: 181393558